requirements.txt
twitchio==2.8.2
openai>=1.0.0
geopy==2.4.1
requests==2.31.0
python-dotenv==1.0.0